Upload existing version of easymasq before cleanup
A basic DNS blocking tool that takes EasyList (AdBlock Plus/UBlock) or DNS Host files and performs syncing on the lists to build a blacklist of domains. This means you can set up cron or other tools to automatically pull down the latest adblocks without the need of any additional software other then wget/curl, sed, tr, and a shell. This is ideal for a router or other embedded systems which might not have the resources to use larger scripts or web interfaces.