add whitelist
added a very basic whitelisting / exemption for any hosts that you want to browse but might get accidentally added into the deny host file because of different list rules. right now its the fourth argument, but i'll have to be switching to optarg in the future to keep the usage clean.